Extracted helper functions
