+++ e2e/cypress/integration/files_and_attachments/file_previews_1_spec.js
@@ -12,7 +12,7 @@
 
 import * as TIMEOUTS from '../../fixtures/timeouts';
 
-import {downloadAttachmentAndVerifyItsProperties} from './helpers';
+import {testImage} from './helpers';
 
 describe('Upload Files - Image', () => {
     let testTeam;
@@ -70,55 +70,3 @@ describe('Upload Files - Image', () => {
         testImage(properties);
     });
 });
-
-function testImage(imageProperties) {
-    const {route, originalWidth, originalHeight} = imageProperties;
-    const filename = route.split('/').pop();
-    const aspectRatio = originalWidth / originalHeight;
-
-    // # Post an image in center channel
-    cy.get('#centerChannelFooter').find('#fileUploadInput').attachFile(route);
-    cy.waitUntil(() => cy.get('#postCreateFooter').then((el) => {
-        return el.find('.post-image__thumbnail').length > 0;
-    }));
-    cy.get('.post-image').should('be.visible');
-    cy.get('#create_post').find('.file-preview').within(() => {
-        // * Img thumbnail exist
-        cy.get('.post-image__thumbnail > div.post-image.normal').should('exist');
-    });
-    cy.postMessage('{enter}');
-    cy.wait(TIMEOUTS.ONE_SEC);
-
-    cy.getLastPost().within(() => {
-        cy.get('.file-view--single').within(() => {
-            // * Image is posted
-            cy.get('img').should('exist').and((img) => {
-            // * Image aspect ratio is maintained
-                expect(img.width() / img.height()).to.be.closeTo(aspectRatio, 0.01);
-            }).click();
-        });
-    });
-
-    cy.get('.modal-body').within(() => {
-        cy.get('.modal-image__content').get('img').should('exist').and((img) => {
-            // * Image aspect ratio is maintained
-            expect(img.width() / img.height()).to.be.closeTo(aspectRatio, 0.01);
-        });
-
-        // # Hover over the image
-        cy.get('.modal-image__content').trigger('mouseover');
-
-        // * Download button should exist
-        cy.findByText('Download').should('exist').parent().then((downloadLink) => {
-            expect(downloadLink.attr('download')).to.equal(filename);
-
-            const fileAttachmentURL = downloadLink.attr('href');
-
-            // * Verify that download link has correct name
-            downloadAttachmentAndVerifyItsProperties(fileAttachmentURL, filename, 'attachment');
-        });
-
-        // # Close modal
-        cy.get('.modal-close').click();
-    });
-}
