+++ e2e/cypress/integration/files_and_attachments/file_previews_4_spec.js
@@ -12,7 +12,7 @@
 
 import * as TIMEOUTS from '../../fixtures/timeouts';
 
-import {downloadAttachmentAndVerifyItsProperties} from './helpers';
+import {testAudioFile} from './helpers';
 
 describe('Upload Files - Audio', () => {
     let testTeam;
@@ -70,51 +70,3 @@ describe('Upload Files - Audio', () => {
         testAudioFile(properties);
     });
 });
-
-function testAudioFile(fileProperties) {
-    const {route, shouldPreview} = fileProperties;
-    const filename = route.split('/').pop();
-
-    // # Post file in center channel
-    cy.get('#centerChannelFooter').find('#fileUploadInput').attachFile(route);
-    cy.waitUntil(() => cy.get('#postCreateFooter').then((el) => {
-        return el.find('.post-image__thumbnail').length > 0;
-    }));
-    cy.get('#create_post').find('.file-preview').within(() => {
-        // * Thumbnail exist
-        cy.get('.post-image__thumbnail > div.audio').should('exist');
-    });
-    cy.postMessage('{enter}');
-    cy.wait(TIMEOUTS.ONE_SEC);
-
-    cy.getLastPost().within(() => {
-        cy.get('.post-image__thumbnail').within(() => {
-            // * File is posted
-            cy.get('.file-icon.audio').should('exist').click();
-        });
-    });
-
-    cy.get('.modal-body').within(() => {
-        if (shouldPreview) {
-            // * Check if the video element exist
-            // Audio is also played by the video element
-            cy.get('video').should('exist');
-        }
-
-        // # Hover over the modal
-        cy.get('.modal-image__content').trigger('mouseover');
-
-        // * Download button should exist
-        cy.findByText('Download').should('exist').parent().then((downloadLink) => {
-            expect(downloadLink.attr('download')).to.equal(filename);
-
-            const fileAttachmentURL = downloadLink.attr('href');
-
-            // * Verify that download link has correct name
-            downloadAttachmentAndVerifyItsProperties(fileAttachmentURL, filename, 'attachment');
-        });
-
-        // # Close modal
-        cy.get('.modal-close').click();
-    });
-}
