+++ e2e/cypress/integration/files_and_attachments/helpers.js
@@ -13,3 +13,193 @@ export function downloadAttachmentAndVerifyItsProperties(fileURL, filename, http
             equal(`${httpContext};filename="${filename}"; filename*=UTF-8''${filename}`);
     });
 }
+
+export function testAudioFile(fileProperties) {
+    const {route, shouldPreview} = fileProperties;
+    const filename = route.split('/').pop();
+
+    // # Post file in center channel
+    cy.get('#centerChannelFooter').find('#fileUploadInput').attachFile(route);
+    cy.waitUntil(() => cy.get('#postCreateFooter').then((el) => {
+        return el.find('.post-image__thumbnail').length > 0;
+    }));
+    cy.get('#create_post').find('.file-preview').within(() => {
+        // * Thumbnail exist
+        cy.get('.post-image__thumbnail > div.audio').should('exist');
+    });
+    cy.postMessage('{enter}');
+    cy.wait(TIMEOUTS.ONE_SEC);
+
+    cy.getLastPost().within(() => {
+        cy.get('.post-image__thumbnail').within(() => {
+            // * File is posted
+            cy.get('.file-icon.audio').should('exist').click();
+        });
+    });
+
+    cy.get('.modal-body').within(() => {
+        if (shouldPreview) {
+            // * Check if the video element exist
+            // Audio is also played by the video element
+            cy.get('video').should('exist');
+        }
+
+        // # Hover over the modal
+        cy.get('.modal-image__content').trigger('mouseover');
+
+        // * Download button should exist
+        cy.findByText('Download').should('exist').parent().then((downloadLink) => {
+            expect(downloadLink.attr('download')).to.equal(filename);
+
+            const fileAttachmentURL = downloadLink.attr('href');
+
+            // * Verify that download link has correct name
+            downloadAttachmentAndVerifyItsProperties(fileAttachmentURL, filename, 'attachment');
+        });
+
+        // # Close modal
+        cy.get('.modal-close').click();
+    });
+}
+
+export function testImage(imageProperties) {
+    const {route, originalWidth, originalHeight} = imageProperties;
+    const filename = route.split('/').pop();
+    const aspectRatio = originalWidth / originalHeight;
+
+    // # Post an image in center channel
+    cy.get('#centerChannelFooter').find('#fileUploadInput').attachFile(route);
+    cy.waitUntil(() => cy.get('#postCreateFooter').then((el) => {
+        return el.find('.post-image__thumbnail').length > 0;
+    }));
+    cy.get('.post-image').should('be.visible');
+    cy.get('#create_post').find('.file-preview').within(() => {
+        // * Img thumbnail exist
+        cy.get('.post-image__thumbnail > div.post-image.normal').should('exist');
+    });
+    cy.postMessage('{enter}');
+    cy.wait(TIMEOUTS.ONE_SEC);
+
+    cy.getLastPost().within(() => {
+        cy.get('.file-view--single').within(() => {
+            // * Image is posted
+            cy.get('img').should('exist').and((img) => {
+            // * Image aspect ratio is maintained
+                expect(img.width() / img.height()).to.be.closeTo(aspectRatio, 0.01);
+            }).click();
+        });
+    });
+
+    cy.get('.modal-body').within(() => {
+        cy.get('.modal-image__content').get('img').should('exist').and((img) => {
+            // * Image aspect ratio is maintained
+            expect(img.width() / img.height()).to.be.closeTo(aspectRatio, 0.01);
+        });
+
+        // # Hover over the image
+        cy.get('.modal-image__content').trigger('mouseover');
+
+        // * Download button should exist
+        cy.findByText('Download').should('exist').parent().then((downloadLink) => {
+            expect(downloadLink.attr('download')).to.equal(filename);
+
+            const fileAttachmentURL = downloadLink.attr('href');
+
+            // * Verify that download link has correct name
+            downloadAttachmentAndVerifyItsProperties(fileAttachmentURL, filename, 'attachment');
+        });
+
+        // # Close modal
+        cy.get('.modal-close').click();
+    });
+}
+
+export function testGenericFile(fileProperties) {
+    const {route, type} = fileProperties;
+    const filename = route.split('/').pop();
+
+    // # Post file in center channel
+    cy.get('#centerChannelFooter').find('#fileUploadInput').attachFile(route);
+    cy.waitUntil(() => cy.get('#postCreateFooter').then((el) => {
+        return el.find('.post-image__thumbnail').length > 0;
+    }));
+    cy.get('#create_post').find('.file-preview').within(() => {
+        // * Thumbnail exist
+        cy.get(`.post-image__thumbnail > div.${type}`).should('exist');
+    });
+    cy.postMessage('{enter}');
+    cy.wait(TIMEOUTS.ONE_SEC);
+
+    cy.getLastPost().within(() => {
+        cy.get('.post-image__thumbnail').within(() => {
+            // * File is posted
+            cy.get(`.file-icon.${type}`).should('exist').click();
+        });
+    });
+
+    cy.get('.modal-body').within(() => {
+        // * No apparent way to check the thumbnail is the correct one.
+        // # Hover over the image
+        cy.get('.modal-image__content').trigger('mouseover');
+
+        // * Download button should exist
+        cy.findByText('Download').should('exist').parent().then((downloadLink) => {
+            expect(downloadLink.attr('download')).to.equal(filename);
+
+            const fileAttachmentURL = downloadLink.attr('href');
+
+            // * Verify that download link has correct name
+            downloadAttachmentAndVerifyItsProperties(fileAttachmentURL, filename, 'attachment');
+        });
+
+        // # Close modal
+        cy.get('.modal-close').click();
+    });
+}
+
+export function testVideoFile(fileProperties) {
+    const {route, shouldPreview} = fileProperties;
+    const filename = route.split('/').pop();
+
+    // # Post file in center channel
+    cy.get('#centerChannelFooter').find('#fileUploadInput').attachFile(route);
+    cy.waitUntil(() => cy.get('#postCreateFooter').then((el) => {
+        return el.find('.post-image__thumbnail').length > 0;
+    }));
+    cy.get('#create_post').find('.file-preview').within(() => {
+        // * Thumbnail exist
+        cy.get('.post-image__thumbnail > div.video').should('exist');
+    });
+    cy.postMessage('{enter}');
+    cy.wait(TIMEOUTS.ONE_SEC);
+
+    cy.getLastPost().within(() => {
+        cy.get('.post-image__thumbnail').within(() => {
+            // * File is posted
+            cy.get('.file-icon.video').should('exist').click();
+        });
+    });
+
+    cy.get('.modal-body').within(() => {
+        if (shouldPreview) {
+            // * Check if the video element exist
+            cy.get('video').should('exist');
+        }
+
+        // # Hover over the image
+        cy.get('.modal-image__content').trigger('mouseover');
+
+        // * Download button should exist
+        cy.findByText('Download').should('exist').parent().then((downloadLink) => {
+            expect(downloadLink.attr('download')).to.equal(filename);
+
+            const fileAttachmentURL = downloadLink.attr('href');
+
+            // * Verify that download link has correct name
+            downloadAttachmentAndVerifyItsProperties(fileAttachmentURL, filename, 'attachment');
+        });
+
+        // # Close modal
+        cy.get('.modal-close').click();
+    });
+}
